translation and fixed project page

+++ src/components/TableRows.tsx
@@ -1,5 +1,6 @@
 import { project } from "../pages/dashboard/utils/interfaces"
 import EditProject from "../pages/projects/components/EditProject"
+import { getId, userRole } from "../util/auth"
 import TableCells, { PropCells } from "./TableCells"
 import { Button } from "./ui/button"
 import { TableCell, TableRow } from "./ui/table"
@@ -16,6 +17,8 @@ export interface Rows {
 }
 
 const TableRows = ({rows,deleteItem,edit}:Rows) =>{
+    console.log(rows)
+    console.log(getId())
 
     return (
         <>
@@ -23,8 +26,8 @@ const TableRows = ({rows,deleteItem,edit}:Rows) =>{
             <TableRow key={row.project.id}>
                 <TableCells data={row.cells}/>
                 <TableCell className="flex space-x-4 justify-center-safe">
-                    {deleteItem && <Button variant="destructive" onClick={()=>deleteItem(row.project.id)}>Delete</Button>}
-                    {edit &&  <EditProject project={row.project} /> }
+                    {(deleteItem && (userRole() === 'ADMIN' || ((userRole() === 'SUPERVISOR')&& getId() == row.project.owner_id))) && <Button variant="destructive" onClick={()=>deleteItem(row.project.id)}>Delete</Button>}
+                    {(edit && (userRole() === 'ADMIN' || ((userRole() === 'SUPERVISOR')&& getId() == row.project.owner_id))) && <EditProject project={row.project} /> }
                 </TableCell>
             </TableRow>)}
         </>
